ListView fixed
Listview now has the desired height. Not sure how it performs with more contents in it, but it should be enough for the purposes of the UI Prototype.

Also, I set the windowSoftInputMode to adjustPan to resolve this weird bug where the software keyboard would shrink the ListView to a ridiculously small size. Not sure if it's perfect this way, but once again, good enough for the Prototype UI.