Options Menu Fully clickable
Some of the functionality is a little dirty, but we really don't have the time to fix that right now so it shouldn't be our focus.

More clearly, the written threads view is a new activity when it really should be just a Fragment. However, ist was kind of complicated starting a new fragment (for me at least) so this it what we're going to have to go with.