Most Client side Login functionality
Client now stores the login credentials in the Shared Preferences if all inputs contain data.
It's not longer possible to leave the input empty an click Login to move forward to the home view, but it still doesn't matter what values are entered in the input field.
Note that the input of "falsch" will no longer force the error message to appear.

Also implemented Utils.java, a class with handy helper methods (for now mostly helpers regarding Shared Preferences).