Login on Service implemented
The LoginActivity checks if a Username/Password pair is valid on the service. If so, it saves the UserID and the returned auth token to the shared preferences.

On the next startup, in the SplashscreenActivity, if there are credentials the Login ist skipped. The validity of the token is not checked yet, but this shouldn't be too much work.

The LoginActivity class is a little untidy and as much as I'd like to write single functions for this, we don't really have the time to do this.