Updated Entryschema
Split up recommendation in counts of positive and negative numbers as we
discussed earlier. This makes it possible for thec client to choose how
he wants to display the rating (e.g. as a percentage or as a bool or as
absolute numbers of up and downvotes).

Also added a field for the parent of an Entry since we will need to know
where to append a comment as a child entry.
This was don according to the PoC (see
https://github.com/Plsr/EISWS1516MaemeckePoplawski/blob/master/MS3/PoC/Source/PoC2/index.js#L9)