Added dummy entry in db
There's a weird behaviour here which I can't quite explain: With a clean
Database (e.g. after Dropping) the dummy entries for course and entry
look fine. However, on the second run, all fields except the _id
disappear. The creating functions are basically the same as for the
dummy user, which works just fine.

For now, the IDs are all I need for testing, but we may need more later
on.