Refactored SingleThreadActivity
Activity now takes an Entry Object in intent extras. This makes it possible for us to display the main thread without an extra request to the service.

Please note, that the only way this is working in this state is after writing a new thread and getting redirected to the SingleThreadActivity.